About REGEX_SMB constant
Today I had a problem while using a password with the + character. The library was throwing the BadMethodCallException located at line 155. So, to avoid it I added the + character in the REGEX_SMB constant. This will allow the connector to work with passwords that have such symbol. It would be a good idea to allow more special characters on passwords and maybe users. I guess a little research would be needed to know which ones should be allowed.